requirements.txt
openai
flask
flask-cors
requests
uvicorn
python-dotenv 
asgiref

